% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{report_sglm}
\alias{report_sglm}
\title{Report ready GLM table with sandwich errors}
\usage{
report_sglm(glm_fit)
}
\arguments{
\item{glm_fit}{glm.fit object for a linear or logistic regression}
}
\value{
tidy data frame with variable, estimate (95% CI), and p-value, ready for reporting
}
\description{
tidy table to obtain ready-to-report estimates and 95% CIs calculated from robust standard errors for a linear or logistic regression model
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
fit <- glm(y~x)
report_sglm(fit)

x <- rnorm(100)
y <- rbinom(100,1,.5)
fit <- glm(y~x, family="binomial")
report_sglm(fit)
}
