% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich_glm}
\alias{sandwich_glm}
\title{GLM table with sandwich errors}
\usage{
sandwich_glm(glm_fit)
}
\arguments{
\item{glm_fit}{glm.fit object for a linear or logistic regression}
}
\value{
tidy tibble for plotting and reporting
}
\description{
tidy table to obtain estimates and 95% CIs calculated from robust standard errors for a linear or logistic regression model
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
fit <- glm(y~x)
sandwich_glm(fit)
}
