% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich_glm_identity}
\alias{sandwich_glm_identity}
\title{identity link sandwich_glm}
\usage{
sandwich_glm_identity(glm_fit)
}
\arguments{
\item{glm_fit}{glm.fit object for a linear regression}
}
\value{
tidy tibble for plotting and reporting
}
\description{
tidy sandwich glm for a linear model (identity link)
}
