% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich_glm_logit}
\alias{sandwich_glm_logit}
\title{logit link sandwich_glm}
\usage{
sandwich_glm_logit(glm_fit)
}
\arguments{
\item{glm_fit}{glm.fit object for a logistic regression}
}
\value{
tidy tibble for plotting and reporting
}
\description{
tidy sandwich glm for a logistic model (logit link)
}
